refactor(1127): utilisation d'une même modale pour les 2 formulaires PE et Milo
